# react-frontend/swagger-gen/swagger.json
{
  "paths" : {
    "/version" : {
      "get" : {
        "operationId" : "version",
        "tags" : [ "routes" ],
        "summary" : "Get Scala Play Version.",
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/models.Version"
            }
          }
        }
      }
    },
    "/color" : {
      "get" : {
        "operationId" : "randomColor",
        "tags" : [ "routes" ],
        "summary" : "Get a random color.",
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/models.Color"
            }
          }
        }
      }
    },
    "/color/{color}" : {
      "get" : {
        "operationId" : "getColor",
        "tags" : [ "routes" ],
        "summary" : "Get a specific color.",
        "parameters" : [ {
          "in" : "path",
          "name" : "color",
          "description" : "Hex color to query.",
          "type" : "string",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/models.Color"
            }
          }
        }
      }
    },
    "/ranking/{limit}" : {
      "get" : {
        "operationId" : "ranking",
        "tags" : [ "routes" ],
        "summary" : "Get the color rankings.",
        "parameters" : [ {
          "in" : "path",
          "format" : "int32",
          "name" : "limit",
          "description" : "Top N to show in ranking.",
          "type" : "integer",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/models.Color"
            }
          }
        }
      }
    },
    "/allColors" : {
      "get" : {
        "operationId" : "allColors",
        "tags" : [ "routes" ],
        "summary" : "Get all colors.",
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/submitChoice" : {
      "post" : {
        "operationId" : "submitChoice",
        "tags" : [ "routes" ],
        "summary" : "Submit color judgement.",
        "parameters" : [ {
          "in" : "query",
          "name" : "color",
          "description" : "Judged color.",
          "type" : "string",
          "required" : true
        }, {
          "in" : "query",
          "name" : "choice",
          "description" : "Color choice for judged color.",
          "type" : "string",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    }
  },
  "produces" : [ "application/json" ],
  "definitions" : {
    "models.Version" : {
      "properties" : {
        "playVersion" : {
          "type" : "string"
        }
      },
      "required" : [ "playVersion" ]
    },
    "models.Color" : {
      "properties" : {
        "hex" : {
          "type" : "string"
        },
        "firstOption" : {
          "type" : "string"
        },
        "secondOption" : {
          "type" : "string"
        },
        "nFirst" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nSecond" : {
          "type" : "integer",
          "format" : "int32"
        }
      },
      "required" : [ "hex", "firstOption", "secondOption", "nFirst", "nSecond" ]
    }
  },
  "swagger" : "2.0",
  "consumes" : [ "application/json" ],
  "info" : {
    "description" : "Supreme Color Truth",
    "title" : "Color Controversy",
    "version" : "1.0-SNAPSHOT"
  },
  "tags" : [ {
    "name" : "routes"
  } ]
}